Separate classes (Box, Line, Board).
Addresses in issue #2.
