+++ single/main.hpp
@@ -38,6 +38,10 @@ using namespace std;
 #define I_LOOP 2000 // 内ループ回数
 #define LIMIT  1800 // 何回以上で不通過マスと判定するか(内ループ)
 
+#include "box.hpp"
+#include "line.hpp"
+#include "board.hpp"
+
 struct Search{
 	int x;
 	int y;
@@ -68,184 +72,6 @@ struct IntraBox{
 //	int nT;	// タッチ数
 //	int nC; // クロス数
 //};
-
-/*******************************************************/
-/** クラス定義 **/
-/*******************************************************/
-
-// ボックス（マス）クラス
-class Box{
-public:
-	enum BoxType {TYPE_NULL,TYPE_NUMBER,TYPE_BLANK};
-	enum FixType {ALL_FIXED,HALF_FIXED,NO_FIXED};
-	static const int NOT_USE = -1;
-	
-	Box(int _x,int _y):type(TYPE_NULL),f_type(NO_FIXED),num(NOT_USE),north(0),south(0),east(0),west(0),f_north(false),f_south(false),f_east(false),f_west(false),self_count(0),candidate(false){x=_x;y=_y;}
-	~Box();
-	
-	bool isTypeNumber() const{return (type==TYPE_NUMBER);}
-	void setTypeNumber(){type=TYPE_NUMBER;}
-	
-	bool isTypeBlank() const{return (type==TYPE_BLANK);}
-	void setTypeBlank(){type=TYPE_BLANK;}
-	
-	bool isTypeAllFixed() const{return (f_type==ALL_FIXED);}
-	void setTypeAllFixed(){f_type=ALL_FIXED;}
-	
-	bool isTypeHalfFixed() const{return (f_type==HALF_FIXED);}
-	void setTypeHalfFixed(){f_type=HALF_FIXED;}
-	
-	int getNumber(){return num;}
-	void setNumber(int _num){num=_num;}
-	
-	int getX(){return x;}
-	int getY(){return y;}
-	
-	void incrementNorthNum(){north++;}
-	void decrementNorthNum(){north--;}
-	int getNorthNum(){return north;}
-	
-	void incrementEastNum(){east++;}
-	void decrementEastNum(){east--;}
-	int getEastNum(){return east;}
-	
-	void incrementSouthNum(){south++;}
-	void decrementSouthNum(){south--;}
-	int getSouthNum(){return south;}
-	
-	void incrementWestNum(){west++;}
-	void decrementWestNum(){west--;}
-	int getWestNum(){return west;}
-	
-	void fixNorthLine(){f_north=true;}
-	void fixEastLine(){f_east=true;}
-	void fixSouthLine(){f_south=true;}
-	void fixWestLine(){f_west=true;}
-	
-	bool isNorthLineFixed(){return f_north;}
-	bool isEastLineFixed(){return f_east;}
-	bool isSouthLineFixed(){return f_south;}
-	bool isWestLineFixed(){return f_west;}
-	
-	void incrementSelfCount(){self_count++;}
-	void resetSelfCount(){self_count=0;}
-	int getSelfCount(){return self_count;}
-	
-	void setCandidate(){candidate=true;}
-	void setNonCandidate(){candidate=false;}
-	bool isCandidate(){return candidate;}
-	
-
-private:
-	BoxType type;	// 数字マスか空白マス
-	FixType f_type; // 固定マス
-	int num;		// 数字マスの数字
-	int x, y;		// 座標
-	int north;		// 北向きに存在する線の数
-	int south;		// 南向きに存在する線の数
-	int east;		// 東向きに存在する線の数
-	int west;		// 西向きに存在する線の数
-	bool f_north;	// 北向きに固定線を持つ？
-	bool f_south;	// 南向きに固定線を持つ？
-	bool f_east;	// 東向きに固定線を持つ？
-	bool f_west;	// 西向きに固定線を持つ？
-	
-	int self_count; // ライン不通過回数カウント用変数
-	bool candidate; // 中間ポートの候補？
-};
-
-// ラインクラス
-class Line{
-public:
-	static const int NOT_USE = -1;
-
-	Line(int _index):x_0(NOT_USE),y_0(NOT_USE),x_1(NOT_USE),y_1(NOT_USE),x_i(NOT_USE),y_i(NOT_USE),candidate(false),im(false){index=_index;track.clear();}
-	~Line();
-
-	void setSourcePort(int x,int y){x_0=x;y_0=y;}
-	int getSourceX(){return x_0;}
-	int getSourceY(){return y_0;}
-	
-	void setSinkPort(int x,int y){x_1=x;y_1=y;}
-	int getSinkX(){return x_1;}
-	int getSinkY(){return y_1;}
-	
-	// 中間ポート
-	void setIntermediatePort(int x,int y){x_i=x;y_i=y;}
-	int getInterX(){return x_i;}
-	int getInterY(){return y_i;}
-	
-	void setCandidate(){candidate=true;}
-	void setNonCandidate(){candidate=false;}
-	bool isCandidate(){return candidate;}
-	
-	void setIntermediateUse(){im=true;}
-	void setIntermediateUnuse(){im=false;}
-	bool isIntermediateUsed(){return im;}
-	
-	void changePort(){	// ソースとシンクの交換
-		int tmp_x = x_0;
-		x_0 = x_1;
-		x_1 = tmp_x;
-		int tmp_y = y_0;
-		y_0 = y_1;
-		y_1 = tmp_y;
-	}
-	
-	int getLineLength(){return track.size();}
-	
-	vector<int>* getTrack(){return &track;}
-	void pushIntToTrack(int t){track.push_back(t);}
-	void clearTrack(){track.clear();}
-
-private:
-	int index;		// ライン番号
-	int x_0, y_0;	// ソースポート
-	int x_1, y_1;	// シンクポート
-	int x_i, y_i;	// 中間ポート
-	bool candidate; // 中間ポートを利用する候補か？
-	bool im;		// 中間ポートを利用するか？
-	vector<int> track;
-};
-
-// ボードクラス
-class Board{
-public:
-	Board(int _x,int _y,int _ln){
-		size_x = _x;
-		size_y = _y;
-		line_num = _ln;
-		for(int y=0;y<size_y;y++){
-			for(int x=0;x<size_x;x++){
-				Box* box;
-				box = new Box(x,y);
-				boxes[y][x] = box;
-			}
-		}
-		for(int i=1;i<=line_num;i++){
-			Line* line;
-			line = new Line(i);
-			lines[i] = line;
-		}
-	}
-	~Board(){
-		// デストラクタ（要メモリ解放）
-	}
-	
-	int getSizeX(){return size_x;}
-	int getSizeY(){return size_y;}
-	int getLineNum(){return line_num;}
-	
-	Box* box(int x,int y){return boxes[y][x];}	// マスの取得
-	Line* line(int idx){return lines[idx];}		// ラインの取得
-	
-private:
-	int size_x;		// X方向サイズ
-	int size_y;		// Y方向サイズ
-	int line_num;	// ライン数
-	map<int,map<int,Box*> > boxes;	// マスの集合
-	map<int,Line*> lines;	// ラインの集合
-};
 
 
 /*******************************************************/
